PLUGINS: Big refactoring
