+++ engines/module.mk
@@ -5,6 +5,7 @@ MODULE_OBJS := \
 	dialogs.o \
 	engine.o \
 	game.o \
+	engineman.o \
 	metaengine.o \
 	obsolete.o \
 	savestate.o
