+++ gui/editgamedialog.cpp
@@ -29,6 +29,8 @@
 #include "common/translation.h"
 #include "common/system.h"
 
+#include "engines/engineman.h"
+
 #include "gui/browser.h"
 #include "gui/gui-manager.h"
 #include "gui/message.h"
@@ -115,7 +117,7 @@ EditGameDialog::EditGameDialog(const String &domain)
 	if (!metaEnginePlugin) {
 		warning("MetaEnginePlugin for target \"%s\" not found!", domain.c_str());
 	} else {
-		enginePlugin = PluginMan.getEngineFromMetaEngine(metaEnginePlugin);
+		enginePlugin = EngineMan.getEngineFromMetaEngine(metaEnginePlugin);
 		if (!enginePlugin) {
 			warning("Engine Plugin for target \"%s\" not found! Game specific settings might be missing.", domain.c_str());
 		}
